Mejoras en la interfaz para depurar

+++ real_rt_estimator/model.h
@@ -149,7 +149,7 @@ namespace real_rt_estimator {
 
 		struct myMatch {
 			int matchNum;
-			int matchDistance;
+			double matchDistance;
 			double matchAprox;
 		};
